+++ apps/server/src/modules/industries/application/use-cases/delete-industry.use-case.ts
@@ -1,0 +1,14 @@
+import { Inject, Injectable } from "@nestjs/common";
+import { INDUSTRY_REPOSITORY, type IndustryRepository } from "../services/industry.repository";
+
+@Injectable()
+export class DeleteIndustryUseCase {
+  public constructor(
+    @Inject(INDUSTRY_REPOSITORY)
+    private readonly industryRepository: IndustryRepository,
+  ) {}
+
+  public execute(industryId: string) {
+    return this.industryRepository.softDelete(industryId);
+  }
+}