+++ apps/server/src/modules/industries/industries.definition.ts
@@ -1,0 +1,6 @@
+import { CxModule } from "@cxnext/core";
+
+export class IndustriesDefinition extends CxModule {
+  public readonly name = "industries";
+  public readonly boundedContext = "organisation";
+}